basic function is nealy completed

+++ src/main/java/com/robben/bbs/mapper/RichTextMapper.java
@@ -1,1 +1,1 @@
-package com.robben.bbs.mapper;import org.springframework.stereotype.Repository;@Repositorypublic interface RichTextMapper {    Integer saveRichText(Integer authorId,String content);    String getRichTextById(Integer id);}
+package com.robben.bbs.mapper;import org.springframework.stereotype.Repository;@Repositorypublic interface RichTextMapper {    Integer saveRichText(Long authorId,String content);    String getRichTextById(Integer id);}
