+++ src/main/java/com/robben/bbs/service/RichTextService.java
@@ -1,1 +1,1 @@
-package com.robben.bbs.service;public interface RichTextService {    Integer saveRichText(Integer authorId,String content);    String getRichTextById(Integer id);}
+package com.robben.bbs.service;public interface RichTextService {    Integer saveRichText(Long authorId,String content);    String getRichTextById(Integer id);}
